+++ src/command/user/login.ts
@@ -1,11 +1,12 @@
-import Login from "../obnizio/login";
-import User from "../obnizio/user";
-import * as Storage from "../storage";
+import { Login } from "../../libs/obnizio/login";
+import User from "../../libs/obnizio/user";
+import * as Storage from "../../libs/storage";
 
-import { getOra } from "../ora-console/getora";
+import { getOra } from "../../libs/ora-console/getora";
+
 const ora = getOra();
 
-export default async (app?: { id: string; token: string }) => {
+export const LoginCommand = async (app?: { id: string; token: string }) => {
   let spinner = ora(`Singin...`).start();
   const token = await Login((text) => {
     spinner.text = text;
