+++ src/command/user/logout.ts
@@ -1,0 +1,6 @@
+import * as Storage from "../../libs/storage";
+
+export const LogoutCommand = async () => {
+  Storage.set("token", null);
+  console.log(`Signed out`);
+};
