--- src/libs/user/logout.ts
@@ -1,6 +1,0 @@
-import * as Storage from "../storage";
-
-export default async () => {
-  Storage.set("token", null);
-  console.log(`Signed out`);
-};
